services/adapter/requirements.txt
numpy==1.26.4
transformers==4.39.3
pymilvus==2.4.0
loguru==0.7.2
amqp==5.2.0
PyYAML==6.0.1
tritonclient[all]==2.44.0
pillow==10.3.0